Refactoring of truncate helper method
I discovered couple of issues with `models.PyCryptoBot.truncate()`
method:

* the precision was not always honored, e.g. `(1.123, 4)` emitted
`"1.123"` which, when printed by logging, broke the alignment of
"columns" in the output:

```
2021-05-24 11:10:00 | ADAUSDT | 5m | Close: 1.4568 | EMA12/26: 1.4568 = 1.4568 | MACD: 0.0000 = 0.0000OBV: 3534629.9200 (0.0000%)ERI: | WAIT
2021-05-24 11:15:00 | ADAUSDT | 5m | Close: 1.458 | *^ EMA12/26: 1.4569 > 1.4568 ^* | *^ MACD: 0.0001 > 0.0000 ^* | ^ OBV: 6328470.5900 (79.0400%) ^ | ERI: buy | WAIT
2021-05-24 11:20:00 | ADAUSDT | 5m | Close: 1.4625 | ^ EMA12/26: 1.4578 > 1.4573 ^ | ^ MACD: 0.0005 > 0.0001 ^ | ^ OBV: 9257233.7600 (46.2800%) ^ | ERI: sell | WAIT
```

(`Close: 1.458` should really be printed as `Close: 1.4580`)

* inconsistent return value: a string, but, with carefuly picked inputs,
a float. AFAICT, every time `truncate()` is called, its return value is
either converted to string (`str(app.truncate(...))`), or to float.
* instance method yet it uses no instance attributes, it's pretty much a
pure function since its output depends only on its inputs \o/

So I took a shot and refactored it to become a standalone function,
fixed the return value to be a string, always - these two steps reduced
quite common `str(app.truncate(...))` to just `truncate(...)`. In the
process, I fixed the issue with precision not being honored, and I added
a unit test to protect it against future regressions.