fix: retrieve all the wallets in the account
The default page size of 50 records may miss the wallet with funds
which results in 'Insufficient funds' error. Currently, Coinbase
list around 70 wallets and using maximum page size of 250 avoid the
funds error. However, this is a temporary fix; until they add more
than 250 wallets. This is an edge case as well and the error is only
seen when a user has selected all the wallets when creating an API
keys.